Move language caching out of Makefile.
This move the project structure just enough to make serving the flask
app packaged and cache the language contents.  This starts to solify on
poetry as the only tool (not ready for several more changes).

This does start establishing omnilingo as the command that controls the
rest of the processing.

This is missing some behavioural testing that the web server actually
does what it should and probably needs to be added so that refactors
like this don't break everything.