Ensure that /index.html and /favicon.ico are requestable.
This adds a test that we can request /index.html and /favicon.ico and
successfully get back a response.  We could also ensure that response
has content or even go as far as to check that the response has no
broken links if desired.

This does not move anything around that is not required to get these
tests to pass.  Thus more things may still be subtly broken as we
continue collecting everything into the omnilingo package.